move mocks -> user/mocks, structs -> user/structs
